requirements.txt
fastapi
uvicorn[standard]
langgraph
langchain-anthropic
langchain-openai
pygithub
pydantic
python-dotenv
httpx